requirements.txt
alembic==1.18.4
aniso8601==10.0.1
attrs==25.4.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-restx==1.3.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
referencing==0.37.0
rpds-py==0.30.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
